chore: fix comments
